Gsn 43 implement authentication check for endpoint access (#13)
* Implement authentication check for user endpoints

* Implement authentication check for post endpoints

* Add get current user endpoint

+++ src/main/java/org/pikovets/GeeksSocialNetworkAPI/service/PostService.java
@@ -3,8 +3,11 @@ package org.pikovets.GeeksSocialNetworkAPI.service;
 import org.pikovets.GeeksSocialNetworkAPI.dto.post.CreatePostRequest;
 import org.pikovets.GeeksSocialNetworkAPI.dto.post.UpdatePostRequest;
 import org.pikovets.GeeksSocialNetworkAPI.exceptions.NotFoundException;
+import org.pikovets.GeeksSocialNetworkAPI.exceptions.UnAuthorizedException;
 import org.pikovets.GeeksSocialNetworkAPI.model.Post;
+import org.pikovets.GeeksSocialNetworkAPI.model.User;
 import org.pikovets.GeeksSocialNetworkAPI.repository.PostRepository;
+import org.pikovets.GeeksSocialNetworkAPI.security.JwtUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,17 +20,20 @@ import java.util.UUID;
 public class PostService {
     private final PostRepository postRepository;
     private final UserService userService;
+    private final JwtUtils jwtUtils;
 
     @Autowired
-    public PostService(PostRepository postRepository, UserService userService) {
+    public PostService(PostRepository postRepository, UserService userService, JwtUtils jwtUtils) {
         this.postRepository = postRepository;
         this.userService = userService;
+        this.jwtUtils = jwtUtils;
     }
 
     @Transactional
-    public void createPost(UUID authorId, CreatePostRequest createRequest) {
+    public void createPost(CreatePostRequest createRequest, String token) {
         Post post = new Post();
-        enrichPost(authorId, post, createRequest);
+        User author = userService.getUserByEmail(jwtUtils.extractUsername(token));
+        enrichPost(author, post, createRequest);
 
         postRepository.save(post);
     }
@@ -37,9 +43,10 @@ public class PostService {
     }
 
     @Transactional
-    public void updatePost(UUID id, UpdatePostRequest updateRequest)
-    {
+    public void updatePost(UUID id, UpdatePostRequest updateRequest, String token) {
         Post postToBeUpdated = postRepository.findById(id).orElseThrow(new NotFoundException("Post not found"));
+
+        validateDeletePermission(postToBeUpdated, token);
 
         Post updatedPost = new Post();
         updatedPost.setText(updateRequest.getText());
@@ -51,16 +58,24 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(UUID id) {
+    public void deletePost(UUID id, String token) {
         Post deletedPost = postRepository.findById(id)
                 .orElseThrow(new NotFoundException("Post not found"));
+
+        validateDeletePermission(deletedPost, token);
 
         postRepository.delete(deletedPost);
     }
 
-    public void enrichPost(UUID authorID, Post post, CreatePostRequest createRequest) {
+    private void enrichPost(User author, Post post, CreatePostRequest createRequest) {
         post.setText(createRequest.getText());
-        post.setAuthor(userService.getUserById(authorID));
+        post.setAuthor(userService.getUserById(author.getId()));
         post.setDate(LocalDateTime.now());
+    }
+
+    private void validateDeletePermission(Post deletedPost, String token) {
+        if (deletedPost.getAuthor().getEmail().equals(jwtUtils.extractUsername(token))) {
+            throw new UnAuthorizedException();
+        }
     }
 }
